Remove ugly chars from end with regex
I noticed in the demo (http://micjamking.github.io/succinct/) that if you use the character count as _30_ it leaves a "," at the end. this bothered me so i made a really simple addition that matches a regex at the end and removes characters like: 

*,.{[(<¦¬~^

this is set as an option so the user can change, remove or whatever. if set to 'false' it wont try to remove.